refactor(backend): split the 1000-line metadata files router
api/metadata/files.py did three jobs. It becomes folders.py (library layout,
fetch-from-downloads), browse.py (tree, paged browsing, filter values) and a
files.py that keeps only per-file read/update/batch/readiness/rules/delete —
1021 lines to 161/457/393.

Shared row-shaping helpers move to _helpers.py. The routers are included in
folders -> browse -> files order so the literal /folders/... paths still win
over /folders/{mode}/...; verified by resolving both against the live app.
